// core/apiClient/api.ts
    export enum StatusEnum {
        PENDING = <any> 'PENDING',
        ACCEPTED = <any> 'ACCEPTED',
        DECLINED = <any> 'DECLINED'
    }
// ...
export interface LatLng {
    /**
     * 
     * @type {number}
     * @memberof LatLng
     */
    latitude: number;
    /**
     * 
     * @type {number}
     * @memberof LatLng
     */
    longitude: number;
}
// ...
export interface ListOfOrganizations {
    /**
     * 
     * @type {Array<Organization>}
     * @memberof ListOfOrganizations
     */
    records?: Array<Organization>;
    /**
     * 
     * @type {number}
     * @memberof ListOfOrganizations
     */
    total?: number;
}
// ...
export interface Organization {
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    address: string;
    /**
     * 
     * @type {number}
     * @memberof Organization
     */
    updatedAt: number;
    /**
     * 
     * @type {number}
     * @memberof Organization
     */
    createdAt: number;
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    createdBy: string;
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    ownerUserProfileId: string;
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    updatedBy: string;
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    id: string;
    /**
     * 
     * @type {LatLng}
     * @memberof Organization
     */
    location: LatLng;
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    name: string;
    /**
     * 
     * @type {boolean}
     * @memberof Organization
     */
    hasActiveSubscription: boolean;
    /**
     * 
     * @type {any}
     * @memberof Organization
     */
    extra?: any;
}
// ...
export interface OrganizationUserProfile {
    /**
     * 
     * @type {string}
     * @memberof OrganizationUserProfile
     */
    userProfileId: string;
    /**
     * 
     * @type {string}
     * @memberof OrganizationUserProfile
     */
    role: OrganizationUserProfile.RoleEnum;
    /**
     * 
     * @type {string}
     * @memberof OrganizationUserProfile
     */
    status: OrganizationUserProfile.StatusEnum;
    /**
     * 
     * @type {any}
     * @memberof OrganizationUserProfile
     */
    extra?: any;
}

/**
 * @export
 * @namespace OrganizationUserProfile
 */
export namespace OrganizationUserProfile {
    /**
     * @export
     * @enum {string}
     */
    export enum RoleEnum {
        ADMIN = <any> 'ADMIN',
        DRIVER = <any> 'DRIVER',
        ASSOCIATE = <any> 'ASSOCIATE'
    }
    /**
     * @export
     * @enum {string}
     */
    export enum StatusEnum {
        UNVERIFIED = <any> 'UNVERIFIED',
        PENDING = <any> 'PENDING',
        ACTIVE = <any> 'ACTIVE',
        DISABLED = <any> 'DISABLED',
        DELETED = <any> 'DELETED'
    }
}
// ...
export interface UserProfile {
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    id: string;
    /**
     * number of milliseconds since epoch
     * @type {number}
     * @memberof UserProfile
     */
    createdAt?: number;
    /**
     * number of milliseconds since epoch
     * @type {number}
     * @memberof UserProfile
     */
    updatedAt?: number;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    createdBy?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    updatedBy?: string;
    /**
     * 
     * @type {any}
     * @memberof UserProfile
     */
    extra?: any;
}
/**
 * 
 * @export
 * @interface UserProfileOrganization
 */
export interface UserProfileOrganization {
    /**
     * 
     * @type {string}
     * @memberof UserProfileOrganization
     */
    organizationId: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfileOrganization
     */
    role?: UserProfileOrganization.RoleEnum;
    /**
     * 
     * @type {string}
     * @memberof UserProfileOrganization
     */
    status?: UserProfileOrganization.StatusEnum;
    /**
     * 
     * @type {any}
     * @memberof UserProfileOrganization
     */
    extra?: any;
}

/**
 * @export
 * @namespace UserProfileOrganization
 */
export namespace UserProfileOrganization {
    /**
     * @export
     * @enum {string}
     */
    export enum RoleEnum {
        ADMIN = <any> 'ADMIN',
        DRIVER = <any> 'DRIVER',
        ASSOCIATE = <any> 'ASSOCIATE'
    }
    /**
     * @export
     * @enum {string}
     */
    export enum StatusEnum {
        UNVERIFIED = <any> 'UNVERIFIED',
        PENDING = <any> 'PENDING',
        ACTIVE = <any> 'ACTIVE',
        DISABLED = <any> 'DISABLED',
        DELETED = <any> 'DELETED'
    }
}